requirements.txt
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
idna==3.7
pycparser==2.22
PyNaCl==1.5.0
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.32.3
urllib3==2.2.2
